Mac made visibile
Mac now is displayed in orange on the main wing and conventional tail
surfaces
Corrected a major error in YCmac along the way :)
If need to make visible on other tailsurfaces add :MACHidden? nil to the
definition of the tailsurface object of the tail object that you wish to
display the MAC on.
The main wing uses a mac object to display it. This comes from the
fundamental difference between wing and tail that we have now
programmed. Future development could improve this...